Added address resolution to mDNS.
The mDNS interface now allows the device to ask who owns a particular
local domain, e.g. 'server.local', by using the new hostQuery() method
in the ESP8266mDNS interface. The implementation for hostQuery()
was mainly due to Erland Lewin, although the input argument for
passing the hostname was modified to accept the entire domain name,
'server.local' instead of just the second level name, 'server'.
This allows nonstandard TLDs for free. Since external domains are
outside the control of the device, its interface should be as
permissive as possible, unlike the other methods in the class.

To be able to use the returned IP address transparently, the SDK's
dns system uses its local host list to lookup cached domains. This
way a simple call to dns_local_addhost() is enough to cache the
local domain address.

Note that it is important to resolve the domain on demand through
the local host list instead of using the IP address directly because
TLS verification matches the domain name used to access the server.
Generating a certificate using an internal IP address would be
rare and fragile.